Optimize chart deletion
Retrieve charts in use using parallelism, rather than serially, before
deleting a given chart. This aims to avoid deletion timeouts when there
is a long list of charts to be retrieved before deletion.

Test Plan:
PASS: Upload and apply 5 StarlingX apps in addition to the default ones.
      Delete those apps.
      Confirm that no timeout errors are raised.
PASS: Upload and apply 5 StarlingX apps in addition to the default ones.
      Update one of the applied apps.
      Confirm that the app updated successfully.

Closes-bug: 2137094

Change-Id: Ie6692cec3e0ef2b9fdb1258b90891df438113f8e
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>